02-use-cases/beginner/a2a_simple/pyproject.toml
[project]
name = "a2a-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "agentkit-sdk-python==0.1.14",
  "google-adk==1.18.0",
  "veadk-python==0.2.29",
]
